extension: Reduce code duplication
