+++ web/packages/core/package.json
@@ -54,5 +54,6 @@
         "ts-node": "^10.9.1",
         "typedoc": "^0.24.0",
         "typescript": "^5.0.0"
-    }
+    },
+    "sideEffects": false
 }
