--- web/packages/core/src/swf-file-name.ts
@@ -1,11 +1,0 @@
-/**
- * Create a filename to save a downloaded SWF into.
- *
- * @param swfUrl The URL of the SWF file.
- * @returns The filename the SWF file can be saved at.
- */
-export function swfFileName(swfUrl: URL): string {
-    const pathName = swfUrl.pathname;
-    const name = pathName.substring(pathName.lastIndexOf("/") + 1);
-    return name;
-}
