+++ web/packages/core/src/swf-utils.ts
@@ -1,0 +1,61 @@
+import {
+    FLASH_MIMETYPE,
+    FUTURESPLASH_MIMETYPE,
+    FLASH7_AND_8_MIMETYPE,
+    FLASH_MOVIE_MIMETYPE,
+} from "./flash-identifiers";
+
+/**
+ * Returns whether the given filename ends in a known flash extension.
+ *
+ * @param filename The filename to test.
+ * @returns True if the filename is a flash movie (swf or spl).
+ */
+export function isSwfFilename(filename: string | null): boolean {
+    if (filename) {
+        let pathname = "";
+        try {
+            // A base URL is required if `filename` is a relative URL, but we don't need to detect the real URL origin.
+            pathname = new URL(filename, "https://example.com").pathname;
+        } catch (err) {
+            // Some invalid filenames, like `///`, could raise a TypeError. Let's fail silently in this situation.
+        }
+        if (pathname && pathname.length >= 4) {
+            const extension = pathname.slice(-4).toLowerCase();
+            if (extension === ".swf" || extension === ".spl") {
+                return true;
+            }
+        }
+    }
+    return false;
+}
+
+/**
+ * Returns whether the given MIME type is a known flash type.
+ *
+ * @param mimeType The MIME type to test.
+ * @returns True if the MIME type is a flash MIME type.
+ */
+export function isSwfMimeType(mimeType: string): boolean {
+    switch (mimeType.toLowerCase()) {
+        case FLASH_MIMETYPE.toLowerCase():
+        case FUTURESPLASH_MIMETYPE.toLowerCase():
+        case FLASH7_AND_8_MIMETYPE.toLowerCase():
+        case FLASH_MOVIE_MIMETYPE.toLowerCase():
+            return true;
+        default:
+            return false;
+    }
+}
+
+/**
+ * Create a filename to save a downloaded SWF into.
+ *
+ * @param swfUrl The URL of the SWF file.
+ * @returns The filename the SWF file can be saved at.
+ */
+export function swfFileName(swfUrl: URL): string {
+    const pathName = swfUrl.pathname;
+    const name = pathName.substring(pathName.lastIndexOf("/") + 1);
+    return name;
+}
